backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.3.1
django-allauth==0.60.1
dj-rest-auth==5.0.2

# Database
psycopg2-binary==2.9.9

# Media Storage (Cloudinary)
cloudinary==1.38.0
django-cloudinary-storage==0.3.0
Pillow==10.2.0

# Utilities
python-decouple==3.8
python-dotenv==1.0.1

# Filters & Search
django-filter==23.5

# Development
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# CORS & Security
django-environ==0.11.2

# API Documentation (optional mais utile)
drf-spectacular==0.27.1
